make the password-clearing algorithm smarter
Remove the ro.config.sim_password_clear property; go back to enabling
this for all devices.  For GSM devices, we clear the password on SIM
change as before.  For CDMA devices, we only clear when the device is
provisioned with a different number; we don't clear the password while
the device is unprovisioned.