Workaround for 2262578: Don't draw tabs if view is not in the correct orientation for the layout
This is an uber hack.  Since there is a race between resizing the view and getting
the orientation-changed notification, this just tries to avoid drawing the tabs in the
wrong orientation (based on what we *expect* to be the orientation specified in the layout
file for the SlidingTab). This masks the problem *most* of the time.