Fix issue #2271668: Live Wallpaper Force-Close when changing from live wallpaper to picasa wallpaper
The detach of the wallpaper engine was coming in after the containing
service was destroyed.  This was possible because these are different
IPC interfaces so will not be serialized.  The fix is to keep track
of all of the active engines in the serice, and detach any remaining
ones when the service is destroyed.

Change-Id: I8810786a777dd4f7b15cfbd2e0da25866a4f3ec5