Finish fixing issue #2228381: android.view.InflateException...
...Binary XML file line #37: Error inflating class <unknown> after adding a secondary account

The problem was that we weren't dealing well with the situation where we start a transition
from activity A to B, then transition back to A before B is shown (it finishes before being
shown), then transition from A to C.  At this point we had some state showing that we
were in the process of showing A from it being hidden (due to the middle transition from
B to A), which would cause the layout pass to ensure its window is hidden before the
transition starts.

The solution is to detect the case where we are showing a token and it is already actually
shown, and in this case not do all of the token setup for it to wait for its windows to
be displayed before it is shown.  This isn't needed, the windows are already displayed
or the token is already set up to wait for them to be displayed.

Change-Id: I16925b91e1e2449dd65ade162a5758173c6e2695