Don't accidentally signal multiple exceptions in setThreadPriority()
After signalling an exception due to set_sched_policy() failing (probably
because the thread in question had exited before this function was called), we
proceeded to call setpriority() anyway, and raised *another* exception when it
also failed.  We don't do this any more; we just bail after signalling the first
exception.

Change-Id: I397b6a93c73d3f0870176e26759f2bec33e56205