Immediately close the shade when the status bar is disabled. (DO NOT MERGE)
This provides a quick fix for http://b/2298803 (if the
in-call UI comes up while the user is holding the
windowshade, the shade sticks in place and the display
becomes unresponsive).

Merged from master (I92a824fe).

TODO: safely restore code to animate the shade closed rather
than simply causing it to disappear.

Change-Id: I53dc76fdfe86636f30f37a223a88d27a5766d3eb