Small oom_adj tweaks.
Include a proper name for processes that are in the background
with running services that have been running for a long time (instead
of showing them as bg-empty).

Batch together multiple processes into the same background bin when
there are significantly more processes than bins.  Arguably this
should be smarter and base the binning on the number of background/empty
processes instead of the total count, but this should be good
enough for now.