At the end of a zoom, only check if the textfield is onscreen.
Previously, we checked to see if the IME was active on the WebTextView,
in an attempt to fix http://b/issue?id=2266066 .  If so, we move the
textfield on screen. However, that is resulting in
http://b/issue?id=2147168 , where a textfield far offscreen is being
scrolled on screen at the end of a zoom, and with this change, the
latter bug is fixed.  Further, I am no longer able to reproduce the
earlier bug with this change.